Remove GNU awk/make/patch/sed wrappers.
These cause more trouble than they're worth, including libtool
configuring in a reference to /usr/libexec/paludis/util, and automake
tests invoking them with PATH=., causing them to fail.  Users will
have to manually arrange for the GNU versions to be visible, on
systems where they are not the default.