Stop overriding PATH in utils
This was apparently added because of sandbox clobbering PATH, but it
doesn't appear to do so anymore, possibly because we avoid executing
sandbox.bashrc now.

Overriding PATH is bad because it means that if util A calls util B,
and util B has a special version for some EAPIs, util A will call the
wrong one unless it to is customised.