Add hack to code with declare statements in environment.bz2
In bash, "declare" executed inside a function magically changes to
"local".  Recentish versions of Portage use "declare" in
environment.bz2, and since we load it back inside a function we need
to work around it.  Thanks to Brian Harring <ferringb@gmail.com> for
reporting and suggesting the fix.

See Gentoo#309369