Don't consider OS ABI when checking library compatibility
glibc accepts both _SYSV and _LINUX, and as of 2.12, creates its own
libraries with _LINUX (and also starts checking the ABI version with a
range rather than always 0) so it doesn't make sense to categorise
files as one or the other as far as mutual compatibility goes.  It's
not clear what the rules should be exactly, so just ignore
unless/until we get complaints/more information.