Fix non-recursive dodoc when $x is a directory
When dodoc is called with a directory there is a chance that it has a
size.

$ mkdir empty; [[ -s empty ]];
$ echo Exited with: $?
Exited with: 0

This means that the case with [[ -s ]]; is taken instead of [[ -d ]];
and that there will be no error set.

This commit fixes it by moving the -d check before the -s check.