cave: add print-unmanaged-files
This adds a new command to cave: print-unmanaged-files.  This command is similar
in spirit to the print-unamanged-files demo ruby script.  The default invocation
will scan the system root and recursively list all files.  You may however limit
it to a subset of the filesystem by specifying the --root parameter.  More than
one root may be specified simultaneously.

Signed-off-by: Saleem Abdulrasool <compnerd@compnerd.org>