stripper: add support to compress DWARF sections
This adds a new dwarf_compress build option to use dwz to compress the DWARF
information in executables and libraries.  Use of this functionality requires
the dwz package to be installed and available in the default path.  Enabling
this during builds can reduce the size of the *compressed* symbols by an
additional ~10-25%.