ebuild: reorder and uniformise PATH
Make the search PATH order prefer the secondary hierarchy (/usr) as it is the
primary point for user binaries.  Still permit falling back to the primary
binary (user and system) directories in the case an executable is not found in
the user supplied path, nor the secondary hierarchy.

Make the PATH setup uniform across normal and binary ebuilds as well as during
VDB updates.