merger: refactor make_arrows
make_arrows is used to generate indicators for FS mergers.  The arrows that are
generated are similar for all FS mergers.  Permit overriding of the arrow
generation in the future if necessary by making the method virtual and hoist it
into the base class FSMerger so that it may be shared across the VDB merger as
well as the NDBAM merger.

Signed-off-by: Saleem Abdulrasool <compnerd@compnerd.org>