build: fix cross compilation check
Target is the code-generation target used for cross-compiling toolchain
components.  Build is the build host, and host is the target host.  Because
paludis is not generating code for a different target, the proper
cross-compilation invocation would set host (and possibly build, although that
is merely for completeness rather than as a need).  Fix the check for
cross-compilation: if the target host does not match the build host, then we are
cross-compiling.

The use of $build and $host should also invoke AC_CANONICAL_{BUILD,HOST} to
ensure that the values are the canonical triples.