test: fix tests with bash 4.3
bash 4.3 changes the semantics of the read builtin.  It will now silently ignore
NUL characters in the input stream.  This is problematic since the paludis pipe
command protocol uses the NUL character as an indicator for end of message.
Passing read an explicit delimiter of a NUL char ($'\0') ensures that it treats
the NUL character as valid input.  This avoids the hang during the process
tests.

The delimiter option has existed prior to 4.3 (and explicitly tested against
4.2p45) and should not break compatibility with older releases of bash.