build: be more lenient about compiler versioning
paludis does not use GNU extensions to C++, use the ISO standard C++11 mode.

Attempt to use the -std=c++11 flag that GNU compatible C++ compilers support.
Fallback to -std=c++0x in case the compiler is older and does not support the
-std=c++11 flag.

This makes the check more resilient to newer compilers as well as different
compilers (e.g. clang).