Don't inappropriately rely on evaluation order
If the key wasn't already present in the map and there were non-trival
default and transformation functions, and the compiler decided to
evaluate the LHS of the assignment before the RHS, the [] would create
the map entry with an empty string value, which would then be returned
by get() and passed to the transformation function instead of using
the default.

Found by compiling with clang (where it broke the PortageEnvironment
tests).