ruby: hand roll ConfigAction
This removes the last user of the EasyAction template in favour of hand expanded
actions.  The utility of the EasyAction template had run out due cross make
InfoAction no longer compatible with it either.  There is a single user of it at
this point.  Just expand it out.  If it becomes useful again in the future, it
can be resurrected.