ruby: use some TMP to reduce redundancy
The BoxedAction and BoxedOptions classes abstract out the details of how to box
and unbox a type across the C++/Ruby boundary.  Traits are used to provide the
type specific information (associating the Action and the Options classes, as
well as the boxed type instance and name spellings).  The result is that the
common patterns are abstracted out as methods which can be materialised as
necessary with compile time checking for cases of methods exposed directly (e.g.
destinations() on certain options classes).