build: modernise use of AM_INIT_AUTOMAKE
The two and three argument forms of AM_INIT_AUTOMAKE are deprecated.  Update the
invocation as recommended.  See:
http://www.gnu.org/software/automake/manual/automake.html#Modernize-AM_005fINIT_005fAUTOMAKE-invocation

In order to keep the previous functionality of only defining the version once,
resort to some acrobatics with m4 macros.  The AC_INIT parameter for the package
version (AC_PACKAGE_VERSION) must be a literal (known at autoconf time).  As a
result, we need to use m4 macros instead of shell values.  The strings cannot be
named the same since the substitution would become value=value rather than
name=value.