PWD is a special variable, we should leave its value alone
Bash sets the value of PWD to the current working directory. It is not a
good idea to change that variable without actually changing the
directory. This would happen if the working directory changes between
saving and loading the environment (saving - when building a PBIN, and
loading - when installing it for example), so better let bash handle
this variable.

Fixes: ticket:1325