add support for cross-compiling
Native cross-compilation is only supported on exheres-0 and paludis-1 capable
environments.  Mark the distributions as supporting cross-compilation.  Use this
to enable the functionality.

Adjust certain repositories to teach paludis that they cannot be cross-compiled
to.

Wire up the cross compile host to various helpers in the resolver to allow the
host to be queried at various times to filter the package set as appropriate.

Add UI to cave to access the cross-compilation support.