utils/unpack: Don't quote UNPACKONLY, it can be interpreted falsely
If $UNPACKONLY is quoted, but is not set, bash sets it to an empty string,
and the tar program will recieve it as an empty argument. On some tar
implementations (bsdtar from libarchive is my test case), this can cause
issues because it'll interpret it as an empty inclusion pattern.

Change-Id: I080db02e983d0ed15a52c9a1152b29a6656cae61
Reviewed-on: https://galileo.mailstation.de/gerrit/2671
Reviewed-by: Bo Ørsted Andresen <zlin@exherbo.org>