Change bash compatibility level handling
Apparently bash is going to deprecate the compat* shopts in favour of
the BASH_COMPAT variable, so use that instead.  BASH_COMPAT is only
supported in 4.3 and later, but it doesn't hurt to set it anyway - in
fact, this is better than the shopt version because 4.2 doesn't
support shopt compat42.  I don't think it's worth messing around with
falling back to shopt for older versions - only EAPIs before 6 would
be affected, and those have survived this long without setting
anything anyway.