support hiding suboptions
Introduce the concept of `SUBOPTIONS_NO_DESCRIBE` to the exheres profile.  This
allows suboptions to be silenced by default from the description list.  This is
useful in rare scenarios, such as `linguas`, where the suboption is common, has
a large set of values.  In such a case, the output of the descriptions would
obscure the actual resolution and useful options.  This new profile variable
permits the profile to indicate to the "frontend" that the description should be
squelched.  Currently, this is consumed only by the cave frontend.