build: unconditionally use visibility flags
Remove the last remnants of icc support as the current released versions do not
support C++11 and we intend to use clang as the alternate compiler.  This cleans
up the compile flags to remove the now obsoleted definition
(-DPALUDIS_USE_VISIBILITY), the unnecessary compile time checks, and the
handling for ICC in the macros.