build: cleanup and simplify flags handling
The ruby headers use some idioms which generate warnings when building paludis.
Rather than assuming that the associated -Wno-* flags are supported by the
compiler frontend, explicitly check for the support when necessary (building the
Ruby bindings).  This slightly speeds up the build when ruby is disabled, but
more importantly, will prevent the accidental usage of these flags more broadly.