build: support out-of-tree builds better
This adjusts the invocation of do_m4.bash to take two parameters, the m4
generator and the destination location.  Doing so permits an out-of-tree build
of paludis when the .m4 files are adjusted and the build needs to re-run the m4
files.  While in the area, fix up a tiny bug with the generation of
comparison_policy.hh.