paludis: POSIX_ME_HARDER accounts repository
POSIX permits an indefinite required buffer size for querying GECOS fields.
This is currently used in three locations.  Previously, paludis would assume
that it could perform a sysconf call to retrieve the requisite buffer size for
ensuring that the subsequent GECOS field query would not receive a -ERANGE
error.  However, as it turns out, this is actually a problem even with GLIBC
where the NSS may end up querying a service which has a larger field value
for the `struct pwd` (glibc sets the return value to the NSS buffer length, aka
1k, but may end up querying a service which provides a larger response).

Use local wrappers which perform the size adjustment to avoid an undersized
buffer.  Localise the functions which query the various GECOS fields which
comprise the persona into util.