util: use `resize` instead of `reserve`
Although `reserve` should be sufficient, it is technically incorrect as per the
spec.  The result of `data` is only valid for the range [data(), data() +
size()].  However, given that a `resize` will not adjust the size of the vector,
the valid range for the data pointer will be the previous capacity, potentially
resulting in a heap corruption.

Thanks to Freundlich for pointing this out.