switch TLS to C++11 TLS semantics
Rather than using the compiler specific `__thread` mechanism for TLS, use the
C++11 standard `thread_local` keyword.  This removes the check for the
`__thread` keyword and the `PALUDIS_TLS` macro in the single usage site.

Take advantage of the fact that `thread_local` can be applied to non-POD types
to avoid the heap allocation of the context buffer and letting the loader
perform the allocation.  This has a slight penalty (an extra function call) on
thread construction.  However, given that the heap allocation would likely be
performed, this will be largely offset by the loss of the allocation.