Ban unprefixed thingies infrastructure.
Adds four new things:

A BANNEDDIR environment variable with location determined by config.
This directory will be prepended to the PATH if set.

A banned_by_distribution util, to make the little stubs survive a
freezing of exheres-0 to exheres-1.

A dobanned that creates little stubs in ${IMAGE}/${BANNEDDIR} that
call banned_by_distribution properly. The stubs are then potentially
shuffled around by the alternatives infrastructure.

BANNEDDIR must not be changed and must not be empty

Change-Id: Ie38703f8922c19a30924a928ddf7e20b24ffa5d0
Reviewed-on: https://galileo.mailstation.de/gerrit/5323
Reviewed-by: Marc-Antoine Perennou <Marc-Antoine@Perennou.com>
Reviewed-by: Bo Ørsted Andresen <zlin@exherbo.org>