run_test.sh: Strip path from test program
TEST_SCRIPT_DIR already contains the path, so it should be removed from
the test program (seems like automake only includes the paths for test
programs with an extension, so TEST_SCRIPT_DIR is still needed for test
programs without extension).

Otherwise out-of-tree builds fail their python tests because the
cleanup/setup scripts aren't found due to the duplicated path:

>>> No ../../../python/../../../python/choices_TEST_setup.sh to run

It turns out that distcheck only worked with automake versions prior to
1.15 because of the folder structure it used. With the duplicated path
it ended up finding the scripts in paludis/python instead of the
distributed ones in paludis/paludis-x.y.z/python.
automake 1.15 changed the build folder for distcheck from $(distdir)/_build
to $(distdir)/_build/sub which caused the distcheck to fail and expose
this problem.

Change-Id: Ie0627bf507d791db7497eeca383160fa50ccbe18
Reviewed-on: https://galileo.mailstation.de/gerrit/5948
Reviewed-by: Bo Ørsted Andresen <zlin@exherbo.org>