build: Work around gcc6&cmake issue #16291 in .gitlab-ci.yaml
A change in gcc 6 made using -isystem with the default include
directory fail. A bug report [1] about this has been closed with
RESOLVED WONTFIX.
This became a problem with cmake[>=3.6] because cmake is filtering
out /usr/include as default include path, but not our non-standard
/usr/${CHOST}/include (upstream issue [2]). Until this is fixed we
can specify the implicit include dir manually. We do the same on
Exherbo in cmake.exlib.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=70129
[2] https://gitlab.kitware.com/cmake/cmake/issues/16291
[3] c077952471b7d8003fb3c41091c2daea01e4f700 in arbor.git