build: Remove redundant checks
The documentation for `cmake_dependent_option` states:
```
Example invocation:

  cmake_dependent_option(USE_FOO "Use Foo" ON "USE_BAR;NOT USE_ZOT" OFF)

If USE_BAR is true and USE_ZOT is false, this provides an option called USE_FOO
that defaults to ON. Otherwise, it sets USE_FOO to OFF and hides the option
from the user. If the status of USE_BAR or USE_ZOT ever changes, any value for
the USE_FOO option is saved so that when the option is re-enabled it retains
its old value.
```

So in this case the `ENABLE_{PYTHON,RUBY}_DOCS` options aren't provided when
`ENABLE_{PYTHON,RUBY}` are disabled and
```
-DENABLE_PYTHON=OFF -DENABLE_PYTHON_DOCS=ON
```
results in `ENABLE_PYTHON=OFF` and `ENABLE_PYTHON_DOCS=OFF`