eapi: implement fetch and mirror overrides.
In EAPI=8, each URI can be prefixed with specific strings to override
mirror or fetch restrictions.

Handle that by checking if URIs start with an override string in the
label handling function adding mirrors.

This isn't very elegant, but works. However, it would be better to add
support for this in the depspec parsing code, add new URI override types
(as... non-leaf nodes, maybe?) and modify the labels accordingly if
overrides has been found, but that would be way more complicated.

Closes: https://github.com/MageSlayer/paludis-gentoo-patches/pull/74