ebuild: support PROPERTIES='test_network'.
According to PMS, this is an optionally supported value and, if set,
overrides a setting of RESTRICT='test', i.e., enables running tests.

This sounds a bit convoluted, and it is, but gets clear with the
rationale behind this property: it is mostly meant as an annotation to
denote that the tests require a network connection to execute and that,
if networking is not available, they will fail. They are, however,
supposed to pass with networking available. To that effect, ebuilds
often disable tests via the RESTRICT feature.

With 'test_network' in the PROPERTIES variable, we can re-enable them
and, additionally, make sure to disable the network sandbox.

Closes: https://github.com/MageSlayer/paludis-gentoo-patches/pull/78