tests: rewrite econf EAPI=8 test in pure sh.
Using constructs like "echo foo | grep ..." opens the door to a lot of
potential pitfalls, since the echo builtin is a real beast. User input
must be specially sanitized to make sure that things work correctly,
which is difficult. Using shell-only features and the printf builtin is
much safer in most circumstances.

Additionally, the old code technically checked for 'datarootdir=...' and
'disable-static' only, but not actually if passed as arguments (i.e.,
missing the dash-dash specifier) for simplicity with grep. By switching
to a for loop, we can just check for the real deal.