tests: don't use illegal suffixes in fetch/mirror restriction override EAPI=8 test.
Doing so would fail the unpacking stage, which is not what we're
interested in.

Strictly speaking, we're fine if this stage fails, since we're only
testing the fetch phase, but it would still be good to not provoke false
negatives.

Also, add a note that this test currently fails with the implementation,
so revisit the implementation code and fix it.