tests: refactor {has{q,v},useq} banning test in EAPI=8 tests.
Instead of testing all functions in one go, split them up into three
packages and test each one individually.

The main reason for this is that the test case has been wrong to begin
with: it's perfectly legal for banned functions to still be defined
according to PMS, the package manager must terminate the build and error
out if they are called, though. We typically ban functions by replacing
them with a wrapper binary (or symlink to such a binary), which is
totally in line with PMS, but couldn't ever pass the original test case.

Instead, just try to call each banned function individually, so that
they should work/return true/0 and modify the test case to check if the
phase fails.

We obviously cannot do this in one ebuild, since we'd only ever test the
first banned function, which is not good enough.