tests: fix and improve assert in subshell test in EAPI=7 tests.
Overriding the test builtin/utility with a function of the same name
doesn't sound like a smart idea, so let's not do this.

Additionally, the ebuild is actually supposed to FAIL if assert works in
a subshell (and some pipe status part is not true), which is what we do,
so reflect that in the test case.