tests: improve dostrip test in EAPI=7 tests.
dostrip might not be a pure shell function, which we probably want to
support, although PMS is a bit unclear in this regard. PMS calls
"docompress" and "dostrip" "functions", but that might just be
negligence.

Additionally, invoke the compiler with the -g3 flag, in order to make
sure that debugging symbols are actually generated. Even then, a very
small subset should be available (like function names), but better make
sure that there is something to really strip.