tests: improve eapply git diff support test in EAPI=7 tests.
The normal git diff syntax is not a huge hurdle, since even older patch
versions could handle that directly (maybe with diagnostic messages, but
still).

The real fun comes in when you consider file renames, copies, permission
changes and symlinks. To that effect, add a patch that tests file
renames.