tests: invalidate the ebuild repository for multiple pretend runs in EAPI=7 tests.
Generating a PackageID with choices (e.g., USE flags) and running the
pretend action on it will also cache it in some way, so that subsequent
generations of the same PackageID retain the choice values internally,
without re-querying them from the environment. Don't ask me for the
specifics of how this is done, I just don't know.

Crucially, though, what that means is that we can't change/override the
choices in the environment and then generate the same PackageID and run
the pretend action on it again, since we'll just be getting the old
choices again.

One (stupid) way around this would be to copy the ebuild (as -r1, -r2
etc.), but that's wasteful and inelegant.

A better way is to just invalidate the ebuild repository's state, which
should, hopefully, not cause other issues.