eapis: create indirections for {,INS}DESTTREE.
These variables were removed with EAPI=7, since they were never meant to
be part of the API in the first place, but crept in.

We remove them by creating indirections and having EAPIs decide what the
variables are supposed to be called. For EAPI=7+, we use different,
hopefully internal names.

Closes: https://github.com/MageSlayer/paludis-gentoo-patches/pull/56