eapis: support optional ENV_UNSET variable.
This does not add the actual implementation yet, only prepares the
internals for its usage.

There might be some confusion about why we need both env_env_unset and
use_env_unset. The first one is supposed to be set to the variable name
that contains such values, while the latter one selects which EAPI
supports this feature.

Unfortunately, we can't just use env_env_unset in EAPI=7+ only and let
it empty for older EAPIs, since it's a profile variable and profile
content is processed before ebuild processing takes place. The profile
as such doesn't have an EAPI (or, rather, it defaults to whatever EAPI
is set as the default if not explicitly provided, which is 0 for Gentoo,
exheres-0 for Exherbo etc.), so the full functionality MUST be available
in the default-if-not-set EAPI.

Hence, we introduce another tools option, use_env_unset, which signifies
whether env_env_unset is actually supposed to be used - hopefully after
the ebuild's EAPI is known.