build: Fix configure check for presence of dirent::d_type
The dirent definitions of both glibc and musl have the dirent::d_type
member, but the configure check fails with:
```
error: 'DT_LINK' undeclared (first use in this function); did you mean 'DT_LNK'?
    6 |     dent.d_type = DT_LINK;
      |                   ^~~~~~~
      |                   DT_LNK
```
since the define is called `DT_LNK` and not `DT_LINK`. Change this to `DT_LNK`
in order to properly detect if we can use dirent::d_type