Replace pcrecpp with std::regex
Upstream stopped supporting the old 'pcre' package in favour of pcre2.
The 8.45 is marked as a final release and nothing else will be
added/fixed in it.

Replace the usage of pcre with `std::regex` that was added in C++11.
This allows us to drop the external dependency on pcre and we no longer
have to provide the functionality via a separate `libcavematchextras`
library that gets dlopen'ed since we no longer link against a
third-party library for it