Disable patsub_replacement for Gentoo EAPIs
Bash 5.2 introduces a new option 'patsub_replacement', which is on by
default and breaks backward compatibility. I added a new config option
to disable it.
Exheres-0 isn't affected because we don't have any BASH_COMPAT there.
This fixes the e_repository test, which started to fail since bash 5.2.

A lot of credit for finding the root cause of the test failure goes to
tombriden.

[1] https://lists.gnu.org/archive/html/bug-bash/2023-02/msg00058.html